Crate opendal_service_ipfs

Apache OpenDAL™ IPFS Service

opendal-service-ipfs provides access to content through IPFS for applications built with Apache OpenDAL™.

Use through opendal

Applications should normally enable this service through the opendal facade with the services-ipfs feature:

cargo add opendal --features services-ipfs

The service is available as opendal::services::Ipfs. Configure the service builder, then pass it to opendal::Operator::new.

Use with opendal-core

Add the split crates directly:

cargo add opendal-core opendal-service-ipfs

Pass a configured service builder to Operator::new:

use opendal_core::{Operator, OperatorRegistry, Result};
use opendal_service_ipfs::{register_ipfs_service, Ipfs};

fn build_operator(builder: Ipfs) -> Result<Operator> {
    Operator::new(builder)
}

fn register_for_uri() {
    register_ipfs_service(OperatorRegistry::get());
}

register_for_uri is only needed for scheme-driven construction through Operator::from_uri or Operator::via_iter.

Services that send HTTP requests also require an HTTP transport in OperationContext. See the opendal-core composition example.

Documentation

Service configuration and examples

Rust API documentation

Apache OpenDAL documentation

License

Licensed under the Apache License, Version 2.0.

Structs

Ipfs

IPFS file system support based on IPFS HTTP Gateway.

IpfsConfig

Config for IPFS file system support.

Constants

IPFS_SCHEME

URI scheme used for service registration and scheme-driven construction.

Functions

register_ipfs_service

Register this service’s URI scheme or schemes with an operator registry.